+++ src/assets/trustAnchorsList.json
